requirements.txt
absl-py==2.3.1
astunparse==1.6.3
attrs==25.3.0
bleach==6.2.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
flatbuffers==25.2.10
fonttools==4.59.1
gast==0.6.0
google-pasta==0.2.0
grpcio==1.74.0
h5py==3.14.0
idna==3.10
jax==0.4.34
jaxlib==0.4.34
joblib==1.5.1
kaggle==1.7.4.5
keras==3.11.3
kiwisolver==1.4.9
libclang==18.1.1
Markdown==3.8.2
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.5
mdurl==0.1.2
mediapipe==0.10.13
ml-dtypes==0.3.2
namex==0.1.0
numpy==1.26.4
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opt_einsum==3.4.0
optree==0.17.0
packaging==25.0
pandas==2.3.2
pillow==11.3.0
protobuf==4.25.3
psutil==7.0.0
pycparser==2.22
Pygments==2.19.2
pyparsing==3.2.3
PyQt6==6.9.1
PyQt6-Qt6==6.9.1
PyQt6_sip==13.10.2
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytz==2025.2
requests==2.32.5
rich==14.1.0
scikit-learn==1.7.1
scipy==1.16.1
seaborn==0.13.2
setuptools==80.9.0
six==1.17.0
sounddevice==0.5.2
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow-cpu==2.16.1
tensorflow-intel==2.16.1
termcolor==3.1.0
text-unidecode==1.3
threadpoolctl==3.6.0
tqdm==4.67.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
webencodings==0.5.1
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.3
pyinstaller==6.15.0